+++ Sources/TDDKit/XCTCaptureError.swift
@@ -7,7 +7,7 @@ public extension XCTestCase {
     /// For example:
     /// ```
     /// func test_failingFetchX_fetchY_fails() async throws {
-    ///     let error = AnyError()
+    ///     let error = XCTError()
     ///     let (sut, _) = makeSUT(fetchXResult: .failure(error))
     ///
     ///     let capturedError = await XCTCaptureError(from: try await sut.fetchY())